Cleaned main.py and extracted scheduler specific methods to helpers.py
